requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
asgiref==3.4.1
astroid==2.9.0
async-timeout==4.0.1
attrs==21.2.0
backoff==1.10.0
botbuilder-core==4.14.0
botbuilder-dialogs==4.14.0
botbuilder-schema==4.14.0
botframework-connector==4.14.0
botframework-streaming==4.14.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==3.4.8
Django==3.2.10
enum34==1.1.10
frozenlist==1.2.0
fyle==0.19.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
isodate==0.6.0
isort==5.10.1
jsonpickle==1.4.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
monotonic==1.6
msal==1.6.0
msrest==0.6.10
multidict==5.2.0
multipledispatch==0.6.0
oauthlib==3.1.1
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==1.7.1
pylint==2.12.1
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2021.3
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3
yarl==1.7.2
